utils: split into {file,string}_utils.{c,h}
