+++ src/lxc/parse.c
@@ -139,55 +139,3 @@ int lxc_file_for_each_line(const char *file, lxc_file_cb callback, void *data)
 	fclose(f);
 	return err;
 }
-
-int lxc_char_left_gc(const char *buffer, size_t len)
-{
-	size_t i;
-
-	for (i = 0; i < len; i++) {
-		if (buffer[i] == ' ' ||
-		    buffer[i] == '\t')
-			continue;
-
-		return i;
-	}
-
-	return 0;
-}
-
-int lxc_char_right_gc(const char *buffer, size_t len)
-{
-	int i;
-
-	for (i = len - 1; i >= 0; i--) {
-		if (buffer[i] == ' '  ||
-		    buffer[i] == '\t' ||
-		    buffer[i] == '\n' ||
-		    buffer[i] == '\0')
-			continue;
-
-		return i + 1;
-	}
-
-	return 0;
-}
-
-char *lxc_trim_whitespace_in_place(char *buffer)
-{
-	buffer += lxc_char_left_gc(buffer, strlen(buffer));
-	buffer[lxc_char_right_gc(buffer, strlen(buffer))] = '\0';
-	return buffer;
-}
-
-int lxc_is_line_empty(const char *line)
-{
-	int i;
-	size_t len = strlen(line);
-
-	for (i = 0; i < len; i++)
-		if (line[i] != ' ' && line[i] != '\t' &&
-		    line[i] != '\n' && line[i] != '\r' &&
-		    line[i] != '\f' && line[i] != '\0')
-			return 0;
-	return 1;
-}
